+++ src/main/java/com/nastech/upmureport/domain/repository/DirRepository.java
@@ -1,5 +1,6 @@
 package com.nastech.upmureport.domain.repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -10,7 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.nastech.upmureport.domain.entity.Dir;
 
 @Transactional
-public interface DirRepository extends JpaRepository<Dir, Integer>{
+public interface DirRepository extends JpaRepository<Dir, BigInteger>{
 	@Query(value = "select * from dir where parentProjId = :parentProjId and deleteFlag = false", nativeQuery = true)
 	List<Dir> findAllByParentProjId(@Param(value = "parentProjId") Integer parentProjId);
 }
